FDTD | Cleanup | Split MPI Python tests
