+++ test/pyWrapper/test_full_system.py
@@ -119,161 +119,6 @@ def test_shieldedPair(tmp_path):
 
 
 @no_mtln_skip
-@no_mpi_skip
-@pytest.mark.mtln
-@pytest.mark.mpi
-@pytest.mark.wires
-@pytest.mark.multiwire
-def test_bundles_mpi_n_ranks(tmp_path):
-    fn = CASES_FOLDER + "mpi/bundles_for_mpi.fdtd.json"
-    solver = FDTD(
-        input_filename=fn,
-        path_to_exe=SEMBA_EXE,
-        mpi_command="mpirun -np 2",
-        run_in_folder=tmp_path,
-    )
-    solver.run()
-    assert solver.hasFinishedSuccessfully()
-
-    # solver = FDTD(input_filename=fn,
-    #               path_to_exe=SEMBA_EXE,
-    #               mpi_command='mpirun -np 3',
-    #               run_in_folder=tmp_path)
-    # solver.cleanUp()
-    # solver.run()
-    # assert solver.hasFinishedSuccessfully()
-
-    # solver = FDTD(input_filename=fn,
-    #               path_to_exe=SEMBA_EXE,
-    #               mpi_command='mpirun -np 4',
-    #               run_in_folder=tmp_path)
-    # solver.cleanUp()
-    # solver.run()
-    # assert solver.hasFinishedSuccessfully()
-
-
-@no_mtln_skip
-@no_mpi_skip
-@pytest.mark.mtln
-@pytest.mark.mpi
-@pytest.mark.wires
-@pytest.mark.multiwire
-def test_bundles_mpi_n_ranks_2(tmp_path):
-    #             (9,10)(10,10)
-    #                        |  (12,12)
-    #           (7,9)   _    |
-    # (1,7)(3,7)    _ _ | |   |
-    #  _ _        | (9,9)|   |
-    # |   | (5,6) |      |   |
-    # |   |_ _    |      |_  |
-    # | (3,6) |   | (10,6) |_|
-    # |       |_ _|    (11,6)(11,4)(12,4)
-    # |    (5,3)(7,3)
-    # |
-    # (1,1)
-    fn = CASES_FOLDER + "mpi/bundles_for_mpi_2.fdtd.json"
-    solver = FDTD(
-        input_filename=fn,
-        path_to_exe=SEMBA_EXE,
-        mpi_command="mpirun -np 2",
-        run_in_folder=tmp_path,
-    )
-    solver.run()
-    assert solver.hasFinishedSuccessfully()
-
-    # solver = FDTD(input_filename=fn,
-    #               path_to_exe=SEMBA_EXE,
-    #               mpi_command='mpirun -np 3',
-    #               run_in_folder=tmp_path)
-    # solver.cleanUp()
-    # solver.run()
-    # assert solver.hasFinishedSuccessfully()
-
-    # solver = FDTD(input_filename=fn,
-    #               path_to_exe=SEMBA_EXE,
-    #               mpi_command='mpirun -np 4',
-    #               run_in_folder=tmp_path)
-    # solver.cleanUp()
-    # solver.run()
-    # assert solver.hasFinishedSuccessfully()
-
-
-@no_mtln_skip
-@no_mpi_skip
-@pytest.mark.mtln
-@pytest.mark.mpi
-@pytest.mark.wires
-@pytest.mark.multiwire
-@pytest.mark.probes
-def test_shieldedPair_mpi(tmp_path):
-    fn = CASES_FOLDER + "shieldedPair/shieldedPair.fdtd.json"
-    solver = FDTD(
-        input_filename=fn,
-        path_to_exe=SEMBA_EXE,
-        mpi_command="mpirun -np 2",
-        run_in_folder=tmp_path,
-    )
-    solver.run()
-
-    probe_files = [
-        "shieldedPair.fdtd_wire_start_line_out_V_75_74_74.dat",
-        "shieldedPair.fdtd_wire_start_line_out_I_75_74_74.dat",
-        "shieldedPair.fdtd_wire_end_line_out_I_75_71_74.dat",
-        "shieldedPair.fdtd_wire_end_line_out_V_75_71_74.dat",
-    ]
-
-    p_expected = []
-    for pf in probe_files:
-        p_expected.append(Probe(OUTPUTS_FOLDER + pf))
-
-    p_solved = []
-    for pf in probe_files:
-        probe_name = "wire_start" if "_wire_start_" in pf else "wire_end"
-        p_solved.append(
-            Probe(_get_solved_probe_filename(solver, probe_name, filename=pf))
-        )
-
-    for i in [0, 3]:
-        solved = np.interp(
-            p_expected[i]["time"].to_numpy(),
-            p_solved[i]["time"].to_numpy(),
-            p_solved[i]["voltage_0"].to_numpy(),
-        )
-        assert np.corrcoef(solved, p_expected[i]["voltage_0"])[0, 1] > 0.999
-        solved = np.interp(
-            p_expected[i]["time"].to_numpy(),
-            p_solved[i]["time"].to_numpy(),
-            p_solved[i]["voltage_1"].to_numpy(),
-        )
-        assert np.corrcoef(solved, p_expected[i]["voltage_1"])[0, 1] > 0.999
-        solved = np.interp(
-            p_expected[i]["time"].to_numpy(),
-            p_solved[i]["time"].to_numpy(),
-            p_solved[i]["voltage_2"].to_numpy(),
-        )
-        assert np.corrcoef(solved, p_expected[i]["voltage_2"])[0, 1] > 0.999
-    for i in [1, 2]:
-        solved = np.interp(
-            p_expected[i]["time"].to_numpy(),
-            p_solved[i]["time"].to_numpy(),
-            p_solved[i]["current_0"].to_numpy(),
-        )
-        assert np.corrcoef(solved, p_expected[i]["current_0"])[0, 1] > 0.999
-        solved = np.interp(
-            p_expected[i]["time"].to_numpy(),
-            p_solved[i]["time"].to_numpy(),
-            p_solved[i]["current_1"].to_numpy(),
-        )
-        assert np.corrcoef(solved, p_expected[i]["current_1"])[0, 1] > 0.999
-        solved = np.interp(
-            p_expected[i]["time"].to_numpy(),
-            p_solved[i]["time"].to_numpy(),
-            p_solved[i]["current_2"].to_numpy(),
-        )
-        assert np.corrcoef(solved, p_expected[i]["current_2"])[0, 1] > 0.999
-
-
-@no_mtln_skip
 @pytest.mark.mtln
 @pytest.mark.wires
 @pytest.mark.dielectric
@@ -381,85 +226,6 @@ def test_holland_short_terminals_match_open_terminals(tmp_path):
 
 
 @no_mtln_skip
-@no_mpi_skip
-@pytest.mark.mtln
-@pytest.mark.mpi
-@pytest.mark.wires
-@pytest.mark.probes
-def test_holland_mtln_mpi(tmp_path):
-    fn = CASES_FOLDER + "holland/holland1981_unshielded.fdtd.json"
-    # no mpi
-    solver = FDTD(input_filename=fn, path_to_exe=SEMBA_EXE, run_in_folder=tmp_path)
-    solver.run()
-    probe_mid_no_mpi = Probe(
-        _get_solved_probe_filename(solver, "mid_point", contains="_I_")
-    )
-
-    # no mpi -np 1
-    solver = FDTD(
-        input_filename=fn,
-        path_to_exe=SEMBA_EXE,
-        mpi_command="mpirun -np 1",
-        run_in_folder=tmp_path,
-    )
-    solver.cleanUp()
-    solver.run()
-    probe_mid_mpi_1 = Probe(
-        _get_solved_probe_filename(solver, "mid_point", contains="_I_")
-    )
-
-    # no mpi -np 2
-    solver = FDTD(
-        input_filename=fn,
-        path_to_exe=SEMBA_EXE,
-        mpi_command="mpirun -np 2",
-        run_in_folder=tmp_path,
-    )
-    solver.cleanUp()
-    solver.run()
-    probe_mid_mpi_2 = Probe(
-        _get_solved_probe_filename(solver, "mid_point", contains="_I_")
-    )
-
-    # # no mpi -np 3
-    # solver = FDTD(input_filename=fn, path_to_exe=SEMBA_EXE,
-    #               flags=['-mtlnwires'], mpi_command='mpirun -np 3',run_in_folder=tmp_path)
-    # solver.cleanUp()
-    # solver.run()
-    # probe_mid_mpi_3 = Probe(_get_solved_probe_filename(
-    #     solver, "mid_point", contains='_I_'))
-
-    expected_f = json.load(
-        open(OUTPUTS_FOLDER + "holland1981_mid_point_expected_current.json")
-    )
-    expected_t, expected_i = np.array([]), np.array([])
-    for data in expected_f["datasetColl"][0]["data"]:
-        expected_t = np.append(expected_t, float(data["value"][0]))
-        expected_i = np.append(expected_i, float(data["value"][1]))
-
-    expected_i_interp = np.interp(
-        probe_mid_no_mpi["time"] - 3.05 * 1e-9, expected_t, expected_i
-    )
-
-    assert np.allclose(
-        expected_i_interp, probe_mid_no_mpi["current_0"], rtol=1e-4, atol=5e-5
-    )
-
-    assert np.allclose(
-        expected_i_interp, probe_mid_mpi_1["current_0"], rtol=1e-4, atol=5e-5
-    )
-
-    assert np.allclose(
-        expected_i_interp, probe_mid_mpi_2["current_0"], rtol=1e-4, atol=5e-5
-    )
-
-    # assert np.allclose(
-    #     expected_i_interp,
-    #     probe_mid_mpi_3['current_0'],
-    #     rtol=1e-4, atol=5e-5)
-
-
-@no_mtln_skip
 @pytest.mark.mtln
 @pytest.mark.wires
 @pytest.mark.multiwire
@@ -490,73 +256,6 @@ def test_unshielded_multiwires(tmp_path):
         p_solved.data["current_1"].to_numpy(),
     )
     assert np.corrcoef(solved_1, p_expected["current_1"])[0, 1] > 0.999
-
-
-@no_mpi_skip
-@pytest.mark.mpi
-@pytest.mark.wires
-@pytest.mark.probes
-def test_towelHanger_mpi(tmp_path):
-    fn = CASES_FOLDER + "towelHanger/towelHanger_mpi.fdtd.json"
-    mpidir = ["x", "y", "z"]
-    for i in range(3):
-        solver = FDTD(
-            input_filename=fn,
-            path_to_exe=SEMBA_EXE,
-            run_in_folder=tmp_path,
-            flags=["-mpidir " + mpidir[i]],
-            mpi_command="mpirun -np 1",
-        )
-        for j in range(len(solver["mesh"]["coordinates"])):
-            cs = [0, 0, 0]
-            cs[0] = solver["mesh"]["coordinates"][j]["relativePosition"][(0 + i) % 3]
-            cs[1] = solver["mesh"]["coordinates"][j]["relativePosition"][(1 + i) % 3]
-            cs[2] = solver["mesh"]["coordinates"][j]["relativePosition"][(2 + i) % 3]
-            solver["mesh"]["coordinates"][j]["relativePosition"] = cs
-
-        intervals = [[[0, 0, 0], [0, 0, 0]]]
-        intervals[0][0][0] = solver["mesh"]["elements"][2]["intervals"][0][0][
-            (0 + i) % 3
-        ]
-        intervals[0][0][1] = solver["mesh"]["elements"][2]["intervals"][0][0][
-            (1 + i) % 3
-        ]
-        intervals[0][0][2] = solver["mesh"]["elements"][2]["intervals"][0][0][
-            (2 + i) % 3
-        ]
-        intervals[0][1][0] = solver["mesh"]["elements"][2]["intervals"][0][1][
-            (0 + i) % 3
-        ]
-        intervals[0][1][1] = solver["mesh"]["elements"][2]["intervals"][0][1][
-            (1 + i) % 3
-        ]
-        intervals[0][1][2] = solver["mesh"]["elements"][2]["intervals"][0][1][
-            (2 + i) % 3
-        ]
-
-        solver["mesh"]["elements"][2]["intervals"] = intervals
-        solver.cleanUp()
-        solver.run()
-
-        p_solved = [
-            Probe(_get_solved_probe_filename(solver, "wire_start")),
-            Probe(_get_solved_probe_filename(solver, "wire_mid")),
-            Probe(_get_solved_probe_filename(solver, "wire_end")),
-        ]
-
-        p_expected = [
-            Probe(OUTPUTS_FOLDER + "towelHanger.fdtd_wire_start_Wz_27_25_30_s1.dat"),
-            Probe(OUTPUTS_FOLDER + "towelHanger.fdtd_wire_mid_Wx_35_25_32_s5.dat"),
-            Probe(OUTPUTS_FOLDER + "towelHanger.fdtd_wire_end_Wz_43_25_30_s4.dat"),
-        ]
-
-        for i in range(3):
-            solved = np.interp(
-                p_expected[i]["time"].to_numpy(),
-                p_solved[i]["time"].to_numpy(),
-                p_solved[i]["current_0"].to_numpy(),
-            )
-            assert np.corrcoef(solved, p_expected[i]["current_0"])[0, 1] > 0.999
 
 
 @pytest.mark.wires
