+++ test/pyWrapper/test_integration.py
@@ -97,24 +97,6 @@ def test_towel_hanger_case_creates_output_probes(tmp_path):
     assert countLinesInFile(probe_end[0]) == 3
 
 
-@no_mpi_skip
-@pytest.mark.mpi
-@pytest.mark.wires
-def test_airplane_case_with_mpi(tmp_path):
-    fn = CASES_FOLDER + "airplane/airplane.fdtd.json"
-    solver = FDTD(
-        fn,
-        path_to_exe=SEMBA_EXE,
-        run_in_folder=tmp_path,
-        flags=["-mapvtk"],
-        mpi_command="mpirun -np 2",
-    )
-    solver.run()
-
-    vtkmapfile = solver.getVTKMap()
-    assert os.path.isfile(vtkmapfile)
-
-
 @pytest.mark.wires
 def test_simple_cabin_initialization(tmp_path):
     fn = CASES_FOLDER + "simple_cabin/simple_cabin.fdtd.json"
@@ -122,24 +104,6 @@ def test_simple_cabin_initialization(tmp_path):
     solver.run()
 
     assert solver.hasFinishedSuccessfully()
-
-
-@no_mpi_skip
-@pytest.mark.mpi
-@pytest.mark.wires
-def test_simple_cabin_initialization_with_mpi(tmp_path):
-    fn = CASES_FOLDER + "simple_cabin/simple_cabin.fdtd.json"
-    solver = FDTD(
-        fn,
-        path_to_exe=SEMBA_EXE,
-        run_in_folder=tmp_path,
-        flags=["-mapvtk"],
-        mpi_command="mpirun -np 2",
-    )
-    solver.run()
-
-    vtkmapfile = solver.getVTKMap()
-    assert os.path.isfile(vtkmapfile)
 
 
 @pytest.mark.probes
